# natours/apis/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Natours API",
    "version": "1.0.0",
    "description": "API for Natours application"
  },
  "servers": [
    {
      "url": "https://natours-rho-self.vercel.app"
    }
  ],
  "paths": {
    "/api/v1/reviews": {
      "get": {
        "summary": "Get all reviews",
        "tags": ["Reviews"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all reviews"
          }
        }
      },
      "post": {
        "summary": "Create a new review",
        "tags": ["Reviews"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully"
          }
        }
      }
    },
...
    "/api/v1/tours": {
      "get": {
        "summary": "Get all tours",
        "tags": ["Tours"],
        "responses": {
          "200": {
            "description": "Successfully retrieved all tours"
          }
        }
      },
      "post": {
        "summary": "Create a new tour",
        "tags": ["Tours"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tour"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tour created successfully"
          }
        }
      }
    },
    "/api/v1/tours/{id}": {
      "get": {
        "summary": "Get a specific tour by ID",
        "tags": ["Tours"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the tour"
          }
        }
      },
      "patch": {
        "summary": "Update a tour",
        "tags": ["Tours"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tour"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tour updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a tour",
        "tags": ["Tours"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tour deleted successfully"
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "summary": "Get all users (Admin only)",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all users"
          }
        }
      },
      "post": {
        "summary": "Create a user (Admin only)",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Review": {
        "type": "object",
        "required": ["review", "rating", "tour", "user"],
        "properties": {
          "review": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "tour": {
            "type": "string",
            "description": "ID of the tour"
          },
          "user": {
            "type": "string",
            "description": "ID of the user"
          }
        }
      },
      "Tour": {
        "type": "object",
        "required": [
          "name",
          "duration",
          "maxGroupSize",
          "difficulty",
          "price",
          "summary",
          "description",
          "imageCover"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "maxGroupSize": {
            "type": "integer"
          },
          "difficulty": {
            "type": "string",
            "enum": ["easy", "medium", "difficult"]
          },
          "ratingsAverage": {
            "type": "number"
          },
          "ratingsQuantity": {
            "type": "integer"
          },
          "price": {
            "type": "number"
          },
          "summary": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageCover": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startDates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "required": ["name", "email", "password", "passwordConfirm"],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["user", "guide", "lead-guide", "admin"]
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "passwordConfirm": {
            "type": "string",
            "format": "password"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}